requirements.txt
jupyter
pandas
geopandas
geopy
xlrd
python-decouple
folium